파일구조 변경, 코드 분할 및 CSS 수정
