+++ src/pages/appointment/_ui/appointment-button-wrap.module.css
@@ -1,0 +1,31 @@
+.appointmentButtonWrap {
+  display: flex;
+  flex-direction: column;
+  margin-top: 24px;
+  gap: 24px;
+}
+
+.appointmentButton {
+  width: 100%;
+  padding: 32px 272px;
+
+  font-size: 38px;
+  font-weight: 500;
+  color: #fe7352;
+  border: solid 1px #fe7352;
+  border-radius: 16px;
+  background-color: #ffffff;
+}
+
+.appointmentButton:hover {
+  color: #ffffff;
+  background-color: #fe7352;
+  transition: all 0.2s ease;
+}
+
+.appointmentButton:disabled {
+  pointer-events: none;
+  border: none;
+  color: #888888;
+  background-color: #d8d8d8;
+}
