+++ src/pages/appointment/_ui/appointment-cancle-modal.module.css
@@ -1,0 +1,89 @@
+.appointmentList {
+  display: flex;
+  flex-direction: column;
+  max-height: 300px;
+  margin-top: 58px;
+  padding: 10px 20px;
+  gap: 10px;
+  overflow-y: auto;
+  border: 3px solid #ff7f50;
+  border-radius: 20px;
+}
+
+.appointmentList {
+  display: flex;
+  flex-direction: column;
+  max-height: 300px;
+  margin-top: 58px;
+  padding: 10px 20px;
+  gap: 15px;
+  overflow-y: auto;
+  border: 3px solid #ff7f50;
+  border-radius: 20px;
+
+  scrollbar-width: none;
+  -ms-overflow-style: none;
+}
+
+.appointmentList::-webkit-scrollbar {
+  display: none;
+}
+
+.myAppointmentWrapper {
+  display: flex;
+}
+
+.myAppointment {
+  width: 370px;
+  margin-right: 10px;
+}
+
+.myAppointment > p {
+  display: flex;
+  color: #888888;
+  font-weight: 700;
+  font-size: 24px;
+}
+
+.modalButtons {
+  display: flex;
+  margin-top: 58px;
+}
+
+.modalButtons > button {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 290px;
+  height: 94px;
+
+  font-size: 48px;
+  font-weight: 700;
+  border-radius: 12px;
+
+  transition: all 0.2s ease;
+  cursor: pointer;
+}
+
+.confirmButton {
+  color: #ff7f50;
+  background-color: #ffffff;
+  border: 2px solid #ff7f50;
+}
+
+.confirmButton:hover {
+  color: #ffffff;
+  background-color: #ff7f50;
+}
+
+.cancelButton {
+  color: #ffffff;
+  background-color: #ff7f50;
+  border: none;
+  border-radius: 10px;
+}
+
+.cancelButton:hover {
+  background-color: #ff6b42;
+  border-color: #ff6b42;
+}
