+++ src/pages/appointment/_ui/appointment-selected-information.jsx
@@ -1,0 +1,36 @@
+import { formattedTime } from '../util';
+import styles from './appointment-selected-information.module.css';
+
+export function SelectedInformation({ appointmentData }) {
+  return (
+    <div className={styles.appointmentInfo}>
+      <h4>선택된 예약 정보</h4>
+      <table className={styles.infoTable}>
+        <tbody>
+          <tr>
+            <td className={styles.infoLabel}>의 료 진</td>
+            <td className={`${styles.infoValue} ${styles.doctorName}`}>
+              {appointmentData.doctor.id
+                ? `${appointmentData.doctor.name} (?두방 내꽈?)`
+                : '선택되지 않음'}
+            </td>
+          </tr>
+          <tr>
+            <td className={styles.infoLabel}>예 약 날 짜</td>
+            <td className={`${styles.infoValue} ${styles.appointmentDate}`}>
+              {appointmentData.date || '선택되지 않음'}
+            </td>
+          </tr>
+          <tr>
+            <td className={styles.infoLabel}>예 약 시 간</td>
+            <td className={`${styles.infoValue} ${styles.appointmentTime}`}>
+              {appointmentData.time
+                ? formattedTime(appointmentData.time)
+                : '선택되지 않음'}
+            </td>
+          </tr>
+        </tbody>
+      </table>
+    </div>
+  );
+}
