+++ src/pages/appointment/_ui/appointment-skeleton-card.module.css
@@ -1,0 +1,59 @@
+.skeletonCard {
+  display: flex;
+  width: fit-content;
+  height: 376px;
+  border: 1px solid #d2d2d2;
+  border-radius: 16px;
+  overflow: hidden;
+}
+
+.skeletonImage {
+  width: 340px;
+  height: 376px;
+  border-radius: 16px 0 0 16px;
+  background: linear-gradient(90deg, #f0f0f0 25%, #e0e0e0 50%, #f0f0f0 75%);
+  background-size: 200% 100%;
+  animation: shimmer 1.5s infinite;
+}
+
+.skeletonCardDescription {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  padding: 50px 21px 39px 22px;
+  gap: 15px;
+}
+
+.skeletonText {
+  height: 45px;
+  border-radius: 15px;
+  background: linear-gradient(90deg, #f0f0f0 25%, #e0e0e0 50%, #f0f0f0 75%);
+  background-size: 200% 100%;
+  animation: shimmer 1.5s infinite;
+}
+
+.skeletonText:first-child {
+  width: 180px;
+}
+
+.skeletonText:nth-child(2) {
+  width: 150px;
+}
+
+.skeletonDepartment {
+  width: 292px;
+  height: 45px;
+  border-radius: 15px;
+  background: linear-gradient(90deg, #f0f0f0 25%, #e0e0e0 50%, #f0f0f0 75%);
+  background-size: 200% 100%;
+  animation: shimmer 1.5s infinite;
+}
+
+@keyframes shimmer {
+  0% {
+    background-position: -200% 0;
+  }
+  100% {
+    background-position: 200% 0;
+  }
+}
