+++ src/pages/appointment/_ui/appointment-skeleton-timetable.jsx
@@ -1,0 +1,11 @@
+import styles from './appointment-skeleton-timetable.module.css';
+
+export function SkeletonTimeTable() {
+  return (
+    <li>
+      <div className={styles.skeletonTimeTable}>
+        <div className={styles.skeletonTimeTableText}></div>
+      </div>
+    </li>
+  );
+}
