+++ src/pages/appointment/_ui/appointment-skeleton-timetable.module.css
@@ -1,0 +1,29 @@
+.skeletonTimeTable {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 385px;
+  height: 70px;
+  border: solid 1px #d2d2d2;
+  border-radius: 16px;
+  background-color: #f5f5f5;
+  cursor: default;
+}
+
+.skeletonTimeTableText {
+  width: 120px;
+  height: 32px;
+  border-radius: 8px;
+  background: linear-gradient(90deg, #f0f0f0 25%, #e0e0e0 50%, #f0f0f0 75%);
+  background-size: 200% 100%;
+  animation: shimmer 1.5s infinite;
+}
+
+@keyframes shimmer {
+  0% {
+    background-position: -200% 0;
+  }
+  100% {
+    background-position: 200% 0;
+  }
+}
