+++ src/pages/appointment/_ui/appointment-timetable.module.css
@@ -1,0 +1,34 @@
+.timeslot > button {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 385px;
+  height: 70px;
+
+  font-size: 32px;
+  color: #888888;
+  border: solid 1px #d2d2d2;
+  border-radius: 16px;
+  background-color: transparent;
+
+  transition: all 0.2s ease;
+}
+
+.timeslot > button:hover {
+  color: #ffffff;
+  border: none;
+  background-color: #fe7352;
+  cursor: pointer;
+}
+
+.timeslot > button:disabled {
+  pointer-events: none;
+  border: none;
+  color: #888888;
+  background-color: #d8d8d8;
+}
+
+.selected {
+  outline: 5px solid #fe7352;
+  outline-offset: 5px;
+}
