+++ src/pages/appointment/appointment-calendar.module.css
@@ -1,0 +1,102 @@
+.srOnly {
+  position: absolute;
+  width: 1px;
+  height: 1px;
+  margin: -1px;
+  padding: 0;
+  overflow: hidden;
+  border: 0;
+  clip: rect(0 0 0 0);
+  clip-path: inset(50%);
+}
+
+.appointmentCalendar {
+  display: flex;
+}
+
+.medicalBookingCalendar {
+  width: 783px !important;
+  height: 700px !important;
+  max-width: none !important;
+  margin-top: 104px;
+}
+
+/* calendar 각 날짜 셀 크기 조절 */
+.medicalBookingCalendar :global(.react-calendar__tile) {
+  width: 80px !important;
+  height: 80px !important;
+  font-size: 30px;
+}
+
+/* calendar 헤더 높이 조절 */
+.medicalBookingCalendar :global(.react-calendar__navigation) {
+  height: 100px !important;
+  font-size: 30px;
+}
+
+/* calendar 요일 헤더 높이 조절 */
+.medicalBookingCalendar :global(.react-calendar__month-view__weekdays) {
+  height: 80px !important;
+  font-size: 30px;
+}
+
+/* 요일 헤더 밑줄 제거 */
+.medicalBookingCalendar :global(abbr[title]) {
+  text-decoration: none !important;
+}
+
+/* 일요일과 토요일 글씨색 빨간색 */
+.medicalBookingCalendar
+  :global(.react-calendar__month-view__weekdays__weekday:first-child),
+.medicalBookingCalendar
+  :global(.react-calendar__month-view__weekdays__weekday:last-child) {
+  color: #d32f2f !important;
+}
+
+/* 오늘 날짜 스타일 */
+.medicalBookingCalendar :global(.react-calendar__tile--now) {
+  background-color: #e3f2fd !important;
+  color: #1976d2 !important;
+  font-weight: 900 !important;
+}
+
+/* 오늘 날짜 호버 스타일 */
+.medicalBookingCalendar :global(.react-calendar__tile--now:hover) {
+  background-color: #bbdefb !important;
+  color: #1976d2 !important;
+}
+
+/* 모든 비활성화된 타일 통일 스타일 */
+.medicalBookingCalendar :global(.react-calendar__tile:disabled) {
+  background-color: #f5f5f5 !important;
+  color: #bdbdbd !important;
+  cursor: not-allowed !important;
+  pointer-events: none !important;
+  opacity: 0.6 !important;
+}
+
+.medicalBookingCalendar :global(.react-calendar__tile:disabled:hover),
+.medicalBookingCalendar :global(.react-calendar__tile:disabled:focus),
+.medicalBookingCalendar :global(.react-calendar__tile:disabled:active) {
+  background-color: #f5f5f5 !important;
+  color: #bdbdbd !important;
+  cursor: not-allowed !important;
+  pointer-events: none !important;
+}
+
+/* 공휴일 비활성화 */
+.medicalBookingCalendar :global(.react-calendar__tile.holiday:disabled) {
+  background-color: #f5f5f5 !important;
+  color: #bdbdbd !important;
+  cursor: not-allowed !important;
+  pointer-events: none !important;
+  opacity: 0.6 !important;
+}
+
+/* 공휴일 호버 효과 제거 */
+.medicalBookingCalendar :global(.react-calendar__tile.holiday:disabled:hover) {
+  background-color: #f5f5f5 !important;
+  color: #bdbdbd !important;
+  cursor: not-allowed !important;
+  pointer-events: none !important;
+}
