+++ src/pages/appointment/appointment-doctor-list.jsx
@@ -1,13 +1,24 @@
-import { DoctorCard } from './appointment-doctor-card';
-import { SkeletonCard } from './appointment-skeleton-card';
-import styles from './appointment-page.module.css';
+import React from 'react';
+import { DoctorCard } from './_ui/appointment-doctor-card';
+import { SkeletonCard } from './_ui/appointment-skeleton-card';
+import styles from './appointment-doctor-list.module.css';
 
-export function DoctorList({
+export const DoctorList = React.memo(function DoctorList({
   doctors,
   isLoading,
   selectedDoctorId,
   onDoctorSelect,
 }) {
+  const skeletonCards = React.useMemo(
+    () => Array.from({ length: 7 }, (_, index) => <SkeletonCard key={index} />),
+    []
+  );
+
+  const createDoctorSelectHandler = React.useCallback(
+    (doctorId, doctorName) => () => onDoctorSelect(doctorId, doctorName),
+    [onDoctorSelect]
+  );
+
   return (
     <section className={styles.doctorList}>
       <h2 className={styles.srOnly}>의료진 목록</h2>
@@ -18,13 +29,11 @@ export function DoctorList({
                 key={doctor.id}
                 doctor={doctor}
                 isSelected={selectedDoctorId === doctor.id}
-                onSelect={() => onDoctorSelect(doctor.id, doctor.name)}
+                onSelect={createDoctorSelectHandler(doctor.id, doctor.name)}
               />
             ))
-          : Array.from({ length: 7 }).map((_, index) => (
-              <SkeletonCard key={index} />
-            ))}
+          : skeletonCards}
       </ul>
     </section>
   );
-}
+});
