+++ src/pages/appointment/appointment-information.module.css
@@ -1,0 +1,8 @@
+.appointmentInformationWapper {
+  float: right;
+}
+
+.appointmentInformation1 {
+  max-width: 783px;
+  margin-left: 40px;
+}
