+++ src/pages/appointment/appointment-modal.module.css
@@ -1,0 +1,34 @@
+.modalOverlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: rgba(0, 0, 0, 0.5);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+}
+
+.modal {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  max-width: 1150px;
+  padding: 110px 278px 100px 278px;
+
+  text-align: center;
+  background: white;
+  border-radius: 20px;
+  border: 3px solid #ff7f50;
+  box-shadow: 0 10px 25px rgba(0, 0, 0, 0.2);
+}
+
+.logoIcon {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 165px;
+  height: 165px;
+}
