+++ src/pages/appointment/appointment-timetable-container.module.css
@@ -1,0 +1,9 @@
+.timeTable {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  margin-top: 34px;
+  padding: 0;
+  gap: 16px;
+
+  list-style: none;
+}
