+++ src/pages/doctors/doctor-detail.jsx
@@ -1,6 +1,5 @@
 import { Link } from 'react-router';
 import styles from '@/pages/doctors/doctor-card.module.css';
-import { Link } from 'react-router';
 
 function DoctorDetail({ doctor }) {
   return (
